Swapping out `detRangeCfg` for `setRange`
Finally found an official LeapMMW document for the HS2xx3A, which makes no mention of `detRangeCfg` at all — not sure why the DFRobot wiki uses it.  It seems to work, though?

By the way, it was never clear to me why there was four detection ranges — it’s not like you can set different sensitivities for them.  Maybe it has something to do with the $JYRPO messages, which I haven’t messed with, so I wouldn’t know.

Regardless, I’m choosing to follow the manufacturer documentation.